requirements.txt
flask==2.2.3
flask-wtf==1.1.1
pymelcloud==2.11.0
requests==2.28.2
python-dotenv==1.0.0
apscheduler==3.10.1
plotly==5.14.0
gunicorn==20.1.0